// lib/types/notifications.d.ts
export type Notification = {
  id: string;
  user_id: string;
  organization_id: string;
  type: string;
  title: string;
  message: string;
  link: string;
  status: string;
  is_read: boolean;
  created_at: string;
  updated_at: string;
};

export type NotificationResponse = {
  data: Notification[];
  cached?: boolean;
  success: boolean;
};